Correct silly copy-paste error
Reviewer1 rightfully complained that 16 um does not make sense.
The value of '16 µm' for day 60 is indeed *not* realistic because it is a copy-paste error from our analysis notebook (search for the second instance of '16' in the [online view] of our analysis notebook, with the last [updated version] on Oct. 1, 2020) and should be '116 µm', which has been corrected in the current version.

[online view]: https://nbviewer.jupyter.org/github/habi/acinar-analysis/blob/master/Analysis.ipynb
[updated version]: https://github.com/habi/acinar-analysis/commit/46b87039819f511cac27613f26f4089fb9bdbc1f