Correct copy-paste error and lingering TeX stuff
We wrongly stated that an alveolus at day 60 has a diameter of 16 um.
This is simply a copy-paste error from the [analysis notebook] where we
get 116 um (search for the second instance of '16' in the version of
Oct 1, 2020).

[analysis notebook]: https://nbviewer.jupyter.org/github/habi/acinar-analysis/blob/master/Analysis.ipynb
[version]: https://github.com/habi/acinar-analysis/commit/46b87039819f511cac27613f26f4089fb9bdbc1f

Also correct some stuff that was still lingering in the Markdown source
from `siunitx`.