Clarify 'how many acini per lung' we chose
Reviewer2 thinks 'it would be helpful to explain how many acini per lung were chosen in the end'.
Let's reference the first table with the exact numbers of acini already
in the methods section.
In the 'stereological analyis' section we state that we analyzed every
third of the extracted acini, randomly starting with the first, second,
or third one.